+++ letsencrypt-auto-source/pieces/bootstrappers/gentoo_common.sh
