+++ letsencrypt-auto-source/pieces/fetch.py
@@ -95,7 +95,7 @@ def verified_new_le_auto(get, tag, temp_dir):
     le_auto_dir = environ.get(
         'LE_AUTO_DIR_TEMPLATE',
         'https://raw.githubusercontent.com/letsencrypt/letsencrypt/%s/'
-        'letsencrypt_auto/') % tag
+        'letsencrypt-auto-source/') % tag
     write(get(le_auto_dir + 'letsencrypt-auto'), temp_dir, 'letsencrypt-auto')
     write(get(le_auto_dir + 'letsencrypt-auto.sig'), temp_dir, 'letsencrypt-auto.sig')
     write(PUBLIC_KEY, temp_dir, 'public_key.pem')
